requirements.txt
requests>=2.32.0
beautifulsoup4>=4.12.2
lxml>=5.2.0
PyYAML>=6.0.2
tqdm>=4.66.0